Remove aragon-sdk service class
