+++ src/services/aragon-sdk/mutations/use-delegate-tokens.ts
@@ -1,8 +1,18 @@
 import {useMutation, UseMutationOptions} from '@tanstack/react-query';
-import {aragonSdkService} from '../aragon-sdk-service';
 import {IDelegateTokensParams} from '../aragon-sdk-service.api';
 import {usePluginClient} from 'hooks/usePluginClient';
-import {DelegateTokensStepValue} from '@aragon/sdk-client';
+import {DelegateTokensStepValue, TokenVotingClient} from '@aragon/sdk-client';
+import {invariant} from 'utils/invariant';
+
+const delegateTokens = async (
+  params: IDelegateTokensParams,
+  client?: TokenVotingClient
+): Promise<AsyncGenerator<DelegateTokensStepValue>> => {
+  invariant(client != null, 'delegateTokens: client is not defined');
+  const data = client.methods.delegateTokens(params);
+
+  return data;
+};
 
 export const useDelegateTokens = (
   options?: UseMutationOptions<
@@ -14,8 +24,7 @@ export const useDelegateTokens = (
   const client = usePluginClient('token-voting.plugin.dao.eth');
 
   return useMutation(
-    (params: IDelegateTokensParams) =>
-      aragonSdkService.delegateTokens(params, client),
+    (params: IDelegateTokensParams) => delegateTokens(params, client),
     options
   );
 };
